chore: initial commit
This sets up the repository for decompiling the US ROM of Pokémon Ranger
2: Shadows of Almia.

This commit provides the ROM fully disassembled with its data file, and
is able to build a matching ROM. The libraries are still hard-coded in
the asm files.

The `crt0.s` file that is linked in the ROM is the one that comes from
the NitroSDK.

The ARM7 binary has not been decompiled, and is included as-is in the
final ROM. As this binary differs from everything that's available at
the time of this commit, the binary used comes directly from a dump of
the ROM, and the `arm7_defs.bin` file has been patched by hand.

The ARM9 binary, DTCM, ITCM and each overlay are in their own asm file.

The data files are included as-is in the final ROM. No effort has been
made yet to unpack them.

Some effort has already been made to identify some of the labels defined
in the asm files. More specifically, library functions, memory
allocation functions and startup functions started to be identified.

Hard-coded addresses were replaced by their corresponding labels. This
has been done automatically, so there might have been some misses.

Sections were put in place for the linker. They should all be present
and correct.

The repository and build system is based off of the one used in the
`pret/pokeplatinum` project.

Signed-off-by: Kuruyia <github@kuruyia.net>