feat: use version strings from librairies
This removes the hard-coded version strings at the beginning of the
`main.s` file, in favor of those that are specified in their respective
librairies.

This allows having a proper `.version` section.

References to those strings were temporarily reverted to hard-coded
addresses, pending the removal of the hard-coded librairies in favor of
linking them properly.

Version strings were fixed in the librairies when it was necessary.

In the linker script, the `.version` section is forced to be kept.
Otherwise, the linker would have removed the whole section has there is
currently no reference to any of the symbols defined there.

Signed-off-by: Kuruyia <github@kuruyia.net>